ImagesController/requirements.txt
boto3>=1.20.0
botocore>=1.23.0
PyYAML>=5.4
docker>=5.0.0
GitPython>=3.1.0
